[bsdf+phase] added unique id to BSDFs and PhaseFunctions and made all BSDFs/Phasefunctions of a scene accessible
Each BSDFs and phase function of a scene are now stored in in separate vector lists of the scene and
are accessible via `Scene::getBSDFs()` and `Scene::getPhaseFunctions()`.
In addition each BSDF/PhaseFuntion now has a unique id for identification. At the end of the scene loading process the uid will be set so that it matches the position of the BSDF/PhaseFunction in the corresponding vector list.
At the same time all string IDs equal "unnamed" will be replaced by "unmamed_bsdf_$UID"/ "unmamed_phase_$UID".
This helps to identify/connect BSDFs later when exploring a scene from outside (i.e., via the python interface).

BSDf: Added access function to the 'diffuseReflectanceTexture', the 'specularReflectanceTexture' and the 'roughnessTexture'.
this also makes it easier to explore BRDF properties from the outside (e.g., when implementing a Mitsuba scene converter)